Denormalise API response structs
Because of how godoc works and IDE completion, embedding another
exported struct results in fields missing in the struct description,
instead of them being inlined. This is rather unhelpful when dealing
with large structs like these, so instead denormalise the lot. This also
simplifies comparing it with generated code from the schema.

This also declutters the API surface when it comes to the number of
exported identifiers.