Add WeatherMeasurePoint
Contrary to other types the JSON schema for this type is missing. It's
been reconstructed using the API response, the docs on the site and the
XML schema.

The only difference between 1.0 and 2.0 appears to be the addition of
the Weather key, based on a diff of the XML schemas.

Fixes: #1