Make Query and composite Filters purely variadic
Though it's nice to use the type system to enforce Query being passed at
least 1 query, and composite Filters to be passed at least 2 filters
using the type system, it results in some rather unergonomic code for
the caller to write as they need to call the function differently with
1, 2 and more than 2 arguments.

In order to make it easier for consumers of this library, these
functions are now only variadic, so you can always splat a slice of
queries or filters in there.

This does mean it's now on the caller to ensure they don't compose an
invalid query, like an AND/OR/NOT/ELEMENTMATCH with only one filter.
This should result in the API returning an error anyway, it just takes
a network round-trip now. We might add a way to validate this client
side later on and return an error when Build() is called.