Make all fields optional
Despite some schemas defining certain elements as having a minOccurs=1,
due to the INCLUDE and EXCLUDE query capabilities it is still possible
for those elements to be missing from the response. As such, we have to
treat all fields as optional, and the only thing we care about is
whether it returns optionally one element or potentially multiple.

This simplifies the xsd parsing, schema and templates a bit as we no
longer need to check for interactions where a field could return a
single element.