Generate types for enums
In case we hit a SimpleType with an enumeration, generate wrapping types
for the enum values.

The wrappers aren't emitted for the internal structs as at that point
we're just parsing the response from the API. They are emitted for the
public structs, so it's easy to check if a method returned a particular
value without needing to know all the string or int constants.